requirements.txt
# Python/Django
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-filter>=23.0,<24.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-phonenumber-field>=7.2,<8.0
phonenumbers>=8.13,<9.0
python-decouple>=3.8,<4.0
django-jazzmin>=3.0
django-debug-toolbar>=4.0

# Database
psycopg2-binary>=2.9,<3.0

# Authentication
pyotp>=2.9,<3.0
qrcode>=7.4,<8.0
fido2>=1.1,<2.0
cryptography>=41.0,<42.0

# GraphQL
graphene>=3.3,<4.0
graphene-django>=3.2,<4.0
cloudinary>=1.38,<2.0
django-cloudinary-storage>=0.3.0
Pillow>=10.0,<11.0

# Background Tasks
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5,<3.0

# WebSockets
channels>=4.0,<5.0
daphne>=4.0,<5.0
channels-redis>=4.1,<5.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Utilities
python-slugify>=8.0,<9.0
Pygments>=2.17,<3.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<5.0
faker>=20.0,<21.0

# Production
sentry-sdk>=0.20.0,<1.0
cryptography>=41.0,<42.0

gunicorn>=21.0,<22.0
whitenoise>=6.5,<7.0

# PDF Generation (for payment receipts)
reportlab>=4.0,<5.0

# Social Authentication
django-allauth>=0.61,<1.0

# Development
ipython>=8.15,<9.0
